bench_models.py refactored
