+++ benchmarks/model_benchmarks/model_create/models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+from ...utils import bench_setup
+
+bench_setup()
+
+
+class Book(models.Model):
+    title = models.CharField(max_length=100)
